mri_service/requirements.txt
fastapi
python-multipart
uvicorn
pillow
tensorflow
numpy 